RESKC-2:fix attachment deletion on copy and hierarchy

+++ coeus-impl/src/main/java/org/kuali/coeus/common/impl/attachment/KcAttachmentDataDaoImpl.java
@@ -20,6 +20,8 @@ package org.kuali.coeus.common.impl.attachment;
 
 import org.apache.commons.lang3.ArrayUtils;
 import org.apache.commons.lang3.StringUtils;
+import org.apache.commons.lang3.builder.EqualsBuilder;
+import org.apache.commons.lang3.builder.HashCodeBuilder;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.kuali.coeus.common.framework.attachment.KcAttachmentDataDao;
@@ -34,6 +36,8 @@ import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
+import java.util.HashSet;
+import java.util.Set;
 import java.util.UUID;
 
 @Component("kcAttachmentDataDao")
@@ -44,6 +48,8 @@ public class KcAttachmentDataDaoImpl implements KcAttachmentDataDao {
     @Autowired
     @Qualifier("dataSource")
     private DataSource dataSource;
+
+    private Set<TableReference> tableReferences;
 
     @Override
     public byte[] getData(String id) {
@@ -81,9 +87,6 @@ public class KcAttachmentDataDaoImpl implements KcAttachmentDataDao {
         }
 
         try (Connection connection = getDataSource().getConnection()) {
-        	if (StringUtils.isNotBlank(id)) {
-        		deleteAttachment(connection, id);
-        	}
         	try (PreparedStatement stmt = connection.prepareStatement("insert into file_data (id, data) values (?, ?)")) {
 	        	String newId = UUID.randomUUID().toString();
 	        	stmt.setString(1, newId);
@@ -93,6 +96,10 @@ public class KcAttachmentDataDaoImpl implements KcAttachmentDataDao {
                 if (LOG.isDebugEnabled()) {
                     LOG.debug("Created attachment data, new id: " + newId);
                 }
+
+            	if (StringUtils.isNotBlank(id)) {
+            		deleteAttachment(connection, id);
+            	}
 
                 return newId;
         	}
@@ -117,10 +124,46 @@ public class KcAttachmentDataDaoImpl implements KcAttachmentDataDao {
     }
         
     protected void deleteAttachment(Connection conn, String id) throws SQLException {
-    	try (PreparedStatement stmt = conn.prepareStatement("delete from file_data where id = ?")) {
-    		stmt.setString(1, id);
-    		stmt.executeUpdate();
+    	if (countReferences(conn, id) == 0) {
+	    	try (PreparedStatement stmt = conn.prepareStatement("delete from file_data where id = ?")) {
+	    		stmt.setString(1, id);
+	    		stmt.executeUpdate();
+	    	}
     	}
+    }
+
+    protected int countReferences(Connection conn, String id) throws SQLException {
+    	if (tableReferences == null) {
+            populateReferences(conn);
+        }
+    	int count = 0;
+    	for (TableReference ref : tableReferences) {
+        	try (PreparedStatement stmt = conn.prepareStatement("select count(*) from "
+        			+ ref.tableName + " where " + ref.columnName + " = ?")) {
+        		stmt.setString(1, id);
+        		try (ResultSet rs = stmt.executeQuery()) {
+        			if (rs.next()) {
+        				count += rs.getInt(1);
+        			}
+        		}
+        	}
+    	}
+    	return count;
+    }
+    
+    protected void populateReferences(Connection conn) throws SQLException {
+    	tableReferences = new HashSet<>();
+        String schema = null;
+        String catalog = conn.getCatalog();
+        if (conn.getMetaData().getSchemas().next()) {
+            schema = conn.getSchema();
+        } else {
+            schema = catalog;
+        }
+        ResultSet rs = conn.getMetaData().getExportedKeys(catalog,schema,"file_data");
+        while (rs.next()) {
+            tableReferences.add(new TableReference(rs.getString("FKTABLE_NAME"), rs.getString("FKCOLUMN_NAME")));
+        }
     }
 
     public DataSource getDataSource() {
@@ -130,4 +173,36 @@ public class KcAttachmentDataDaoImpl implements KcAttachmentDataDao {
     public void setDataSource(DataSource dataSource) {
         this.dataSource = dataSource;
     }
+    
+    class TableReference {
+    	public String tableName;
+		public String columnName;
+    	public TableReference(String tableName, String columnName) {
+			this.tableName = tableName;
+			this.columnName = columnName;
+		}
+    	@Override
+		public int hashCode() {
+    		return new HashCodeBuilder(17, 37).append(tableName).append(columnName).toHashCode();
+		}
+		@Override
+		public boolean equals(Object obj) {
+			if (this == obj)
+				return true;
+			if (obj == null)
+				return false;
+			if (getClass() != obj.getClass())
+				return false;
+			final TableReference other = (TableReference) obj;
+			return new EqualsBuilder().append(tableName, other.tableName).append(columnName, other.columnName).isEquals();
+		}
+    }
+
+	public Set<TableReference> getTableReferences() {
+		return tableReferences;
+	}
+
+	public void setTableReferences(Set<TableReference> tableReferences) {
+		this.tableReferences = tableReferences;
+	}
 }
